Split symbol resolution into multiple files
